+++ CHANGELOG.md
@@ -7,7 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
----
+### Changed
+
+- **Breaking**: Rename collection to `upcloud.cloud` to enable releasing to Ansible Galaxy.
 
 ## [0.6.0] - 2023-09-25