+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.7.0] - 2025-06-24
+
 ### Changed
 
 - **Breaking**: Rename collection to `upcloud.cloud` to enable releasing to Ansible Galaxy.
@@ -35,7 +37,8 @@ First release for the new UpCloud Ansible Collection! :tada:
 
 - Ansible inventory for UpCloud servers, allowing filtering with zones, tags, networks or server states.
 
-[Unreleased]: https://github.com/UpCloudLtd/upcloud-ansible-collection/compare/v0.6.0...HEAD
+[Unreleased]: https://github.com/UpCloudLtd/upcloud-ansible-collection/compare/v0.7.0...HEAD
+[0.7.0]: https://github.com/UpCloudLtd/upcloud-ansible-collection/releases/tag/v0.7.0
 [0.6.0]: https://github.com/UpCloudLtd/upcloud-ansible-collection/releases/tag/v0.6.0
 [0.5.1]: https://github.com/UpCloudLtd/upcloud-ansible-collection/releases/tag/v0.5.1
 [0.5.0]: https://github.com/UpCloudLtd/upcloud-ansible-collection/releases/tag/v0.5.0